ArgumentParser workarounds
- Due to how the parser do NOT track where the extras came from, we need
  to trace this manually based on out of band information that we have
  from the construction of the parser.

This basically removes the confusing case where end-users do this::

    $ calmjs npm --install calmjs.dev -v -v
    ...
    usage: calmjs [-h] [-v] [-q] [-d] <command> ...
    calmjs: error: unrecognized arguments: -v -v

Where the -v is an accepted argument, however really not accepted by the
subcommand.  In our case this gives an even further misleading due to
how the verbose output seemingly is processed, but this was set up by
the bootstrap parser which sets the logging up.  Perhaps all subparsers
will need to inherit those options.

Well, at the very least, with this fix the correct output will be
produced like so::

    $ calmjs npm --init calmjs.dev --unknown-option
    usage: calmjs npm [-h] [--view] [--init] [--install] [-i] [-m] [-w]
                      package_names [package_names ...]
    calmjs npm: error: unrecognized arguments: --unknown-option

Rather than giving out the root usage like::

    usage: calmjs [-h] [-v] [-q] [-d] <command> ...
    calmjs: error: unrecognized arguments: --unknown-option